% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPred.R
\name{doPred}
\alias{doPred}
\title{doPred}
\usage{
doPred(object, newdata, verbose = FALSE, s)
}
\arguments{
\item{object}{A fitted object of class \code{hal}}

\item{newdata}{A matrix of new predictions to obtain predictions}

\item{verbose}{A \code{boolean} indicating whether to print output on functions progress}

\item{s}{Option from \code{glmnet} indicating what value to use. Defaults to the value
that was specified in the original call to \code{hal} (which itself defaults to the value
which minimizes MSE).}

\item{...}{Additional arguments (not currently used)}
}
\description{
A helper function for the predict method for class \code{hal}.
}
