% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSparseMat.R
\name{getIntersect}
\alias{getIntersect}
\title{getIntersect}
\usage{
getIntersect(...)
}
\arguments{
\item{...}{Arguments passed to \code{lapply}}
}
\description{
Heper function for higher order interaction basis functions.
}
