% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSparseMat.R
\name{makeSparseMat}
\alias{makeSparseMat}
\title{makeSparseMat}
\usage{
makeSparseMat(X, newX = X, verbose = TRUE)
}
\arguments{
\item{X}{A \code{data.frame} of predictors}

\item{newX}{Optional \code{data.frame} on which to return predicted values}

\item{verbose}{A \code{boolean} indicating whether to print output on functions progress}
}
\description{
Function to create a sparse design matrix of basis functions
based on a matrix of predictors.
}
