% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSparseMat.R
\name{myIntersect}
\alias{myIntersect}
\title{myIntersect}
\usage{
myIntersect(...)
}
\arguments{
\item{...}{Arguments passed to intersect}
}
\description{
Helper function for higher order interaction basis functions.
}
