% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hal.R
\name{predict.hal}
\alias{predict.hal}
\title{predict.hal}
\usage{
\method{predict}{hal}(object, newdata, bigDesign = FALSE, verbose = TRUE,
  chunks = 1000, s = ifelse(object$useMin, object$object$lambda.min,
  object$object$lambda.1se), ...)
}
\arguments{
\item{object}{A fitted object of class \code{hal}}

\item{newdata}{A matrix of new predictions to obtain predictions}

\item{bigDesign}{A boolean indicating whether to obtain predictions all at once
(which may be memory intractable) or to split up the task into smaller chunks}

\item{chunks}{A numeric indicating how many chunks to split the prediction task into
(if \code{bigDesign = FALSE})}

\item{s}{Option from \code{glmnet} indicating what value to use. Defaults to the value
that was specified in the original call to \code{hal} (which itself defaults to the value
which minimizes MSE).}

\item{...}{Additional arguments (not currently used)}
}
\description{
A predict method for objects of class \code{hal}.
}
