% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.hal.R, R/predict.SL.hal.R
\name{predict.SL.hal}
\alias{predict.SL.hal}
\alias{predict.SL.hal}
\title{predict.SL.hal}
\usage{
\method{predict}{SL.hal}(object, newdata, bigDesign = FALSE, verbose = TRUE,
  chunks = 1000, s = ifelse(object$useMin, object$object$lambda.min,
  object$object$lambda.1se), ...)

\method{predict}{SL.hal}(object, newdata, bigDesign = FALSE, verbose = TRUE,
  chunks = 1000, s = ifelse(object$useMin, object$object$lambda.min,
  object$object$lambda.1se), ...)
}
\arguments{
\item{object}{A fitted object of class \code{hal}}

\item{newdata}{A matrix of new predictions to obtain predictions}

\item{bigDesign}{A boolean indicating whether to obtain predictions all at once
(which may be memory intractable) or to split up the task into smaller chunks}

\item{...}{Other arguments passed to \code{predict}}

\item{nChunks}{A numeric indicating how many chunks to split the prediction task into
(if \code{bigDesign = FALSE})}
}
\description{
predict method for \code{SL.hal}

Quick pass-through.
}
