% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.hal.R, R/SL.hal_new.R
\name{SL.hal}
\alias{SL.hal}
\alias{SL.hal}
\title{Super Learner wrapper for the Highly Adaptive LASSO}
\usage{
SL.hal(Y, X, newX, family = gaussian(), verbose = TRUE,
  obsWeights = rep(1, length(Y)), sparseMat = TRUE,
  nfolds = ifelse(length(Y) <= 100, 20, 10), nlambda = 100,
  minVars = NULL, maxDim = 10, useMin = TRUE, ...)

SL.hal(Y, X, newX, family = gaussian(), verbose = TRUE,
  obsWeights = rep(1, length(Y)), sparseMat = TRUE,
  nfolds = ifelse(length(Y) <= 100, 20, 10), nlambda = 100,
  minVars = NULL, maxDim = 10, useMin = TRUE, ...)
}
\arguments{
\item{Y}{A \code{numeric} of outcomes}

\item{X}{A \code{data.frame} of predictors}

\item{newX}{Optional \code{data.frame} on which to return predicted values}

\item{family}{Needs to have a character object in \code{family$family} as
required by \code{SuperLearner}}

\item{verbose}{A \code{boolean} indicating whether to print output on the
progress of the function}

\item{obsWeights}{Optional \code{vector} of observation weights to be passed
to \code{cv.glmnet}}

\item{sparseMat}{Use sparse matrix implementation or normal matrix implementation.}

\item{nfolds}{Number of cross-validation folds; passed to \code{cv.glmnet}}

\item{nlambda}{Number of lambda values to search across in \code{cv.glmnet}}

\item{useMin}{Option passed to \code{cv.glmnet}: use minimum risk lambda or
1se lambda (more penalization)}

\item{...}{Any other arguments to pass-through to hal()}

\item{Y}{outcome}

\item{X}{data}

\item{newX}{New data to apply the model fit and generate predictions.}

\item{family}{Statistical family}

\item{verbose}{Set to T for more detailed output}

\item{obsWeights}{observation weights}

\item{nfolds}{Number of CV folds for cv.glmnet}

\item{nlambda}{Number of lambda values to test in cv.glmnet.}

\item{useMin}{Glmnet option - use minimum risk lambda or 1se lambda (more
penalization).}

\item{...}{Any other arguments to pass-through to \code{hal}}
}
\description{
\code{SuperLearner} wrapper for \code{hal}.

SuperLearner wrapper for the highly adaptive LASSO
}
