% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{hal}
\alias{hal}
\title{Highly Adaptive Lasso}
\usage{
hal(Y, X, newX = NULL, family = gaussian(), verbose = F,
  obsWeights = rep(1, length(Y)), sparseMat = TRUE,
  nfolds = ifelse(length(Y) <= 100, 20, 10), nlambda = 100, useMin = TRUE)
}
\arguments{
\item{Y}{outcome}

\item{X}{data}

\item{newX}{New data to apply the model fit and generate predictions.}

\item{family}{Statistical family}

\item{verbose}{Set to T for more detailed output}

\item{obsWeights}{observation weights}

\item{sparseMat}{Use sparse matrix implementation or normal matrix implementation.}

\item{nfolds}{Number of CV folds for cv.glmnet}

\item{nlambda}{Number of lambda values to test in cv.glmnet.}

\item{useMin}{Glmnet option - use minimum risk lambda or 1se lambda (more
penalization).}
}
\description{
Model fitting function.
}

