% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSparseMat.R
\name{makeSparseMat}
\alias{makeSparseMat}
\title{Creates a sparseMatrix of the indicators}
\usage{
makeSparseMat(X, newX = X, verbose = TRUE)
}
\arguments{
\item{X}{X}

\item{newX}{newX}

\item{verbose}{If true output extra info during execution.}
}
\description{
Add more details.
}
