% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.SL.hal.R
\name{predict.SL.hal}
\alias{predict.SL.hal}
\title{Quick pass-through.}
\usage{
\method{predict}{SL.hal}(object, newdata, bigDesign = FALSE, verbose = TRUE,
  chunks = 1000, s = ifelse(object$useMin, object$object$lambda.min,
  object$object$lambda.1se), ...)
}
\description{
Quick pass-through.
}
