% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.hal.R
\name{predict.SL.hal}
\alias{predict.SL.hal}
\title{predict.SL.hal}
\usage{
\method{predict}{SL.hal}(object, newdata, bigDesign = FALSE, chunks = 5000,
  ...)
}
\arguments{
\item{object}{A fitted object of class \code{hal}}

\item{newdata}{A matrix of new predictions to obtain predictions}

\item{bigDesign}{A boolean indicating whether to obtain predictions all at once
(which may be memory intractable) or to split up the task into smaller chunks}

\item{chunks}{A numeric indicating how many observations to use in each chunk of
the prediction task (if \code{bigDesign = FALSE})}

\item{...}{Other arguments passed to \code{predict}}
}
\description{
predict method for \code{SL.hal}
}
