% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.hal.R, R/predict.SL.hal.R
\name{predict.SL.hal}
\alias{predict.SL.hal}
\alias{predict.SL.hal}
\title{Prediction method for the HAL SuperLearner wrapper}
\usage{
\method{predict}{SL.hal}(object, newdata, bigDesign = FALSE, verbose = TRUE,
  chunks = 1000, s = ifelse(object$useMin, object$object$lambda.min,
  object$object$lambda.1se), ...)

\method{predict}{SL.hal}(object, newdata, bigDesign = FALSE, verbose = TRUE,
  chunks = 1000, s = ifelse(object$useMin, object$object$lambda.min,
  object$object$lambda.1se), ...)
}
\arguments{
\item{object}{A fitted object of class \code{hal}}

\item{newdata}{A matrix of new predictions on which to obtain predictions}

\item{bigDesign}{A boolean indicating whether to obtain predictions all at
once (which may be memory intensive) or to split up the
task into smaller chunks}

\item{chunks}{A numeric indicating how many observations to use in each
chunk of the prediction task (if \code{bigDesign = FALSE})}

\item{...}{Other arguments passed to \code{predict}}
}
\description{
method for \code{predict} for \code{SL.hal}

Quick pass-through.
}
