% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.hal.R
\name{SL.hal}
\alias{SL.hal}
\title{Highly Adaptive Lasso}
\usage{
SL.hal(Y, X, newX, family = gaussian(), verbose = TRUE,
  obsWeights = rep(1, length(Y)), sparseMat = TRUE,
  nfolds = ifelse(length(Y) <= 100, 20, 10), nlambda = 100, useMin = TRUE,
  ...)
}
\arguments{
\item{Y}{outcome}

\item{X}{data}
}
\description{
SuperLearner wrapper
}

