% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.hal.R
\name{SL.hal}
\alias{SL.hal}
\title{Highly Adaptive Lasso}
\usage{
SL.hal(Y, X, newX, family = gaussian(), verbose = TRUE,
  obsWeights = rep(1, length(Y)), nfolds = ifelse(length(Y) <= 100, 20, 10),
  nlambda = 100, useMin = TRUE, ...)
}
\arguments{
\item{Y}{A \code{numeric} of outcomes}

\item{X}{A \code{data.frame} of predictors}

\item{newX}{Optional \code{data.frame} on which to return predicted values}

\item{verbose}{A \code{boolean} indicating whether to print output on functions progress}

\item{obsWeights}{Optional \code{vector} of observation weights to be passed to \code{cv.glmnet}}

\item{nfolds}{Number of CV folds passed to \code{cv.glmnet}}

\item{nlambda}{Number of lambda values to search across in \code{cv.glmnet}}

\item{useMin}{Option passed to \code{cv.glmnet}, use minimum risk lambda or 1se lambda (more
penalization)}

\item{...}{Any other arguments to pass-through to hal()}
}
\description{
SuperLearner wrapper for \code{hal}.
}
